refactor: 엔티티 패키지 구조 변경 및 코드 정리
- BaseEntity.java의 패키지를 변경
- Channel.java의 패키지를 변경
- Message.java의 패키지를 변경
- Receivable.java의 패키지를 변경
- User.java의 패키지를 변경
