+++ discordit/src/main/java/com/sprint/mission/discodeit/entity/BaseEntity.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.discodeit.entity.base;
+package com.sprint.mission.discodeit.entity;
 
 import lombok.Getter;
