+++ discordit/src/main/java/com/sprint/mission/discodeit/entity/Message.java
@@ -1,6 +1,5 @@
-package com.sprint.mission.discodeit.entity.base;
+package com.sprint.mission.discodeit.entity;
 
-import com.sprint.mission.discodeit.entity.BinaryContent;
 import com.sprint.mission.discodeit.enums.ReceiverType;
 import lombok.Getter;
 
@@ -41,7 +40,8 @@ public class Message extends BaseEntity {
 
     // 더미 데이터 생성 or 파일 데이터 복원용 (createdAt, updateAt 조작 위해 엔터티 내 배치.)
     public static Message fromDto(UUID uuid, Instant createdAt, Instant updatedAt,
-                                  User sender, ReceiverType type, Receivable receiver, String message, List<BinaryContent> contents) {
+                                  User sender, ReceiverType type, Receivable receiver,
+                                  String message, List<BinaryContent> contents) {
         Message msg = new Message(sender, type, receiver, message);
         msg.attachments.addAll(contents);
         msg.uuid = uuid;
